- tb size unit display at search dialog added
git-svn-id: svn://svn.code.sf.net/p/cdcat/svn/trunk@246 23e97676-e3e9-41ff-bed1-4a4288a26c68

+++ src/find.h
@@ -97,8 +97,8 @@ private:
     QDateTime dateStart;
     QDateTime dateEnd;
 
-    int size_min;
-    int size_max;
+    float size_min;
+    float size_max;
 };
 
 /***********************************************************************/
